Allow to choose Dark/Light tray icon (#1474)
* Allow to choose Dark/Light tray icon

Closes #1356

* Unify light/dark tray icon switching

Before this change, the app displayed a dark tray icon on Linux and a
red one on Windows, regardless of the OS's dark mode setting. On macOS,
it used the OS mechanism via tray image names *Template.png, allowing
macOS to automatically invert the tray icon colors.

With this change, the app now relies on the Electron API: nativeTheme.
This works on macOS, Linux, and Windows. When the OS 'Dark Mode' setting
is enabled, the app will choose a white tray icon; otherwise, it will
use a black one.

Additionally, there is a new setting that allows the user to force the
tray icon to be either black or white. This is useful in situations
where a light theme is in use, but the tray panel is still black, which
is the default configuration on Ubuntu, for example.

* Use black and white logos for tray icon on Linux

* Use the colored tray icon on Windows

* Use trayTemplate.png as tray icon on macOS

* Remove unnecessary tray icon files

* Show the tray icon style setting only on Linux

* Add CSS for "Tray icon style" setting

* Use `select` component for the tray icon style

* Use numbers for `trayIconStyle` in config